+++ code/crates/engine/src/consensus.rs
@@ -1010,7 +1010,7 @@ where
             Effect::SignProposal(proposal, r) => {
                 let start = Instant::now();
 
-                let signed_proposal = self.signing_provider.sign_proposal(proposal);
+                let signed_proposal = self.signing_provider.sign_proposal(proposal).await;
 
                 self.metrics
                     .signature_signing_time
@@ -1022,7 +1022,7 @@ where
             Effect::SignVote(vote, r) => {
                 let start = Instant::now();
 
-                let signed_vote = self.signing_provider.sign_vote(vote);
+                let signed_vote = self.signing_provider.sign_vote(vote).await;
 
                 self.metrics
                     .signature_signing_time
@@ -1040,10 +1040,12 @@ where
                     Msg::Vote(v) => {
                         self.signing_provider
                             .verify_signed_vote(&v, &msg.signature, &pk)
+                            .await
                     }
                     Msg::Proposal(p) => {
                         self.signing_provider
                             .verify_signed_proposal(&p, &msg.signature, &pk)
+                            .await
                     }
                 };
 
@@ -1055,41 +1057,36 @@ where
             }
 
             Effect::VerifyCommitCertificate(certificate, validator_set, thresholds, r) => {
-                let result = self.signing_provider.verify_commit_certificate(
-                    &self.ctx,
-                    &certificate,
-                    &validator_set,
-                    thresholds,
-                );
+                let result = self
+                    .signing_provider
+                    .verify_commit_certificate(&self.ctx, &certificate, &validator_set, thresholds)
+                    .await;
 
                 Ok(r.resume_with(result))
             }
 
             Effect::VerifyPolkaCertificate(certificate, validator_set, thresholds, r) => {
-                let result = self.signing_provider.verify_polka_certificate(
-                    &self.ctx,
-                    &certificate,
-                    &validator_set,
-                    thresholds,
-                );
+                let result = self
+                    .signing_provider
+                    .verify_polka_certificate(&self.ctx, &certificate, &validator_set, thresholds)
+                    .await;
 
                 Ok(r.resume_with(result))
             }
 
             Effect::VerifyRoundCertificate(certificate, validator_set, thresholds, r) => {
-                let result = self.signing_provider.verify_round_certificate(
-                    &self.ctx,
-                    &certificate,
-                    &validator_set,
-                    thresholds,
-                );
+                let result = self
+                    .signing_provider
+                    .verify_round_certificate(&self.ctx, &certificate, &validator_set, thresholds)
+                    .await;
 
                 Ok(r.resume_with(result))
             }
 
             Effect::ExtendVote(height, round, value_id, r) => {
                 if let Some(extension) = self.extend_vote(height, round, value_id).await? {
-                    let signed_extension = self.signing_provider.sign_vote_extension(extension);
+                    let signed_extension =
+                        self.signing_provider.sign_vote_extension(extension).await;
                     Ok(r.resume_with(Some(signed_extension)))
                 } else {
                     Ok(r.resume_with(None))
@@ -1097,11 +1094,14 @@ where
             }
 
             Effect::VerifyVoteExtension(height, round, value_id, signed_extension, pk, r) => {
-                let valid = self.signing_provider.verify_signed_vote_extension(
-                    &signed_extension.message,
-                    &signed_extension.signature,
-                    &pk,
-                );
+                let valid = self
+                    .signing_provider
+                    .verify_signed_vote_extension(
+                        &signed_extension.message,
+                        &signed_extension.signature,
+                        &pk,
+                    )
+                    .await;
 
                 if !valid {
                     return Ok(r.resume_with(Err(VoteExtensionError::InvalidSignature)));