+++ code/crates/test/tests/unit/certificates/commit.rs
@@ -1,3 +1,4 @@
+use futures::executor::block_on;
 use malachitebft_core_types::CommitCertificate;
 
 use super::{make_validators, types::*, CertificateBuilder, CertificateTest, DEFAULT_SEED};
@@ -24,7 +25,12 @@ impl CertificateBuilder for Commit {
         validator_set: &ValidatorSet,
         threshold_params: ThresholdParams,
     ) -> Result<(), CertificateError<TestContext>> {
-        signer.verify_commit_certificate(ctx, certificate, validator_set, threshold_params)
+        block_on(signer.verify_commit_certificate(
+            ctx,
+            certificate,
+            validator_set,
+            threshold_params,
+        ))
     }
 }